# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "backend",
    "description": "",
    "contact": {
      "name": "David Hewitt",
      "email": "davidmhewitt@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/users/github/callback": {
      "get": {
        "tags": [
          "users::github_callback"
        ],
        "operationId": "github_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": ""
          }
        }
      }
    },
...
    "/users/login": {
      "post": {
        "tags": [
          "users::login"
        ],
        "operationId": "login_user",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "303": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginUser": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
...
    }
  }
}